Refactor purchase creation screen
