+++ lib/features/purchases/presentation/models/reception_status_choice.dart
@@ -1,0 +1,2 @@
+/// Defines the reception status for a purchase.
+enum ReceptionStatusChoice { toReceive, alreadyReceived }
